updated footer components to hold everything
